remove LegacyPhoneNumberNormalizerTrait
