+++ src/Serializer/Normalizer/PhoneNumberNormalizer.php
@@ -11,6 +11,8 @@
 
 namespace Misd\PhoneNumberBundle\Serializer\Normalizer;
 
+use Symfony\Component\Serializer\Exception\InvalidArgumentException;
+use Symfony\Component\Serializer\Exception\UnexpectedValueException;
 use Symfony\Component\Serializer\Normalizer\ArrayDenormalizer;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
@@ -31,6 +33,42 @@ if (class_exists(ArrayDenormalizer::class) && !method_exists(ArrayDenormalizer::
      */
     class PhoneNumberNormalizer implements NormalizerInterface, DenormalizerInterface
     {
-        use LegacyPhoneNumberNormalizerTrait;
+        use CommonPhoneNumberNormalizerTrait;
+
+        /**
+         * {@inheritdoc}
+         *
+         * @throws InvalidArgumentException
+         */
+        public function normalize($object, $format = null, array $context = [])
+        {
+            return $this->doNormalize($object, $format, $context);
+        }
+
+        /**
+         * {@inheritdoc}
+         */
+        public function supportsNormalization($data, $format = null)
+        {
+            return $this->doSupportsNormalization($data, $format);
+        }
+
+        /**
+         * {@inheritdoc}
+         *
+         * @throws UnexpectedValueException
+         */
+        public function denormalize($data, $class, $format = null, array $context = [])
+        {
+            return $this->doDenormalize($data, $class, $format, $context);
+        }
+
+        /**
+         * {@inheritdoc}
+         */
+        public function supportsDenormalization($data, $type, $format = null)
+        {
+            return $this->doSupportsDenormalization($data, $type, $format);
+        }
     }
 }
